separate interface tests to test to offer it to all adapter
