+++ tests/interface-tests/package.json
@@ -1,11 +1,18 @@
 {
-	"name": "@loom-io/test-utils",
-	"version": "1.0.0",
+	"name": "@loom-io/interface-tests",
+	"version": "0.6.0",
 	"description": "",
 	"main": "index.js",
 	"type": "module",
 	"scripts": {
-		"build": "tsc -p tsconfig.json"
+		"clean": "rm -rf dist",
+		"build": "pnpm clean && tsc -p tsconfig.json"
+	},
+	"devDependencies": {
+		"@loom-io/core": "workspace:*"
+	},
+	"peerDependencies": {
+		"vitest": "1.3.1"
 	},
 	"exports": {
 		".": {
